Create library for world generation
A library for world generation and rendering was created. It consists of
two classes that represent a column/chunk of voxel blocks. The chunk
class will be able to generate a renderable mesh utilizing the
BlockChunkRenderable class.

API adjustments where made on the graphics library to accommodate for
the needs of the world map.