Add chunk mesh generation
The world generation was extended with the generation of the visible
chunk mesh.

A Renderable class was created in the graphics library to handle a tasks
to be done for the mesh generation and rendering of an arbitrary object.
To render an object a class has to be inherited from Renderable and also
provide it with information in form of a struct about the vertex data.