Add texture class and entity-component-system
The graphics library was extended with a texture class which handles -
obviously - textures. These will be read by the stb-image library, added
as a git submodule.

A entity-component-system for block and chunk data was created to
capsule block data and render ability from one another. It is extendable
with physics and network components for the future.

Lastly, a config parser was created which will parse for example block
type data and build up the internal block data base.